+++ gradle/libs.versions.toml
@@ -166,7 +166,6 @@ maplibre = "org.maplibre.gl:android-sdk:10.2.0"
 maplibre_ktx = "org.maplibre.gl:android-sdk-ktx-v7:2.0.1"
 maplibre_annotation = "org.maplibre.gl:android-plugin-annotation-v9:2.0.1"
 opusencoder = "io.element.android:opusencoder:1.1.0"
-audiowaveform = "com.github.lincollincol:compose-audiowaveform:1.1.1"
 
 # Analytics
 posthog = "com.posthog.android:posthog:2.0.3"
