+++ settings.gradle.kts
@@ -35,7 +35,6 @@ dependencyResolutionManagement {
             content {
                 includeModule("com.github.UnifiedPush", "android-connector")
                 includeModule("com.github.matrix-org", "matrix-analytics-events")
-                includeModule("com.github.lincollincol", "compose-audiowaveform")
             }
         }
         // To have immediate access to Rust SDK versions
